collection_examples

This is a small data frame of species information for use with the shipping/transferal notes as well as with the label makers.

Usage

collection_examples

Format

A data frame with some collections which will one day be shipped.

Details

this data frame is derived from collections in 2023, which have been cleaned with the BarnebyLives pipeline

Examples

head(collection_examples)
#> # A tibble: 6 × 70
#>   Collection_number Primary_Collector     Associated_Collectors Genus    Epithet
#>               <dbl> <chr>                 <chr>                 <chr>    <chr>  
#> 1              2819 Reed Clark Benkendorf NA                    Poa      secunda
#> 2              2820 Reed Clark Benkendorf NA                    Prunus   anders…
#> 3              2821 Reed Clark Benkendorf NA                    Layia    glandu…
#> 4              2822 Reed Clark Benkendorf NA                    Phlox    longif…
#> 5              2823 Reed Clark Benkendorf NA                    Lomatium nevade…
#> 6              2824 Reed Clark Benkendorf NA                    Camisso… contor…
#> # ℹ 65 more variables: Full_name <chr>, Name_authority <chr>,
#> #   Binomial_authority <chr>, Binomial_authority_issues <chr>, Infrarank <chr>,
#> #   Infraspecies <chr>, Infraspecific_authority <chr>, Infra_auth_issues <chr>,
#> #   Family <chr>, Project_name <chr>, Site_name <chr>, latitude_dd <dbl>,
#> #   longitude_dd <dbl>, Datum <chr>, Directions <chr>, Vegetation <chr>,
#> #   Associates <chr>, Habitat <chr>, Fide <chr>, Determined_by <chr>,
#> #   Aspect <chr>, Slope <dbl>, Notes <chr>, Tissue_collections <dbl>, …