take mdy format date and make it written herbarium label format

take the 'month day year' date format, popular in North America, and place it into a museum compliant text format

Usage

date2text(x)

Arguments

x

a data frame with dates

Examples

first50dates <- paste0(sample(3:9, size = 50, replace = TRUE), '-',
   sample(1:29, size = 50, replace = TRUE), '-',
   rep(2023, times = 50 ))
head(first50dates)
#> [1] "9-19-2023" "9-25-2023" "6-2-2023"  "4-8-2023"  "5-1-2023"  "3-28-2023"
first50dates <- date2text(first50dates)
head(first50dates)
#> [1] "19 Sep, 2023" "25 Sep, 2023" "2 Jun, 2023"  "8 Apr, 2023"  "1 May, 2023" 
#> [6] "28 Mar, 2023"