herbaria_info

This is a small data frame of herbaria contact info

Usage

herbaria_info

Format

Each is a data frame with columns relevant to letter heads

Details

this data.frame data set is derived from Index Herbariorum and features some places many of our collectors send specimens to.

Examples

head(herbaria_info)
#>   Abbreviation         Contact                      Name1
#> 1           US   Erika Gardner    U.S. National Herbarium
#> 2         RENO     Jerry Tiehm                  Herbarium
#> 3          UTC Kristian Valles    Intermountain Herbarium
#> 4          OSC   James Mickley                  Herbarium
#> 5         CHIC    Nyree Zerega Nancy Rich Poole Herbarium
#> 6         NSDA     Jeff Knight                  Herbarium
#>                                                       Name2    State
#> 1                             Smithsonian Institution, NMNH         
#> 2 Department of Natural Resources and Environmental Science   Nevada
#> 3                                        College of Science     Utah
#> 4                     Botany and Plant Pathology Department   Oregon
#> 5                         Plant Conservation Science Center Illinois
#> 6                            Nevada Division of Agriculture   Nevada
#>               City    ZipCode                       Street
#> 1 Washington, D.C.      20560 10th & Constitution Ave., NW
#> 2             Reno      89557 1664 N. Virginia St., MS-186
#> 3            Logan 84322-5305           5305 Old Main Hill
#> 4        Corvallis      97331           2701 SW Campus Way
#> 5          Glencoe      60022          1000 Lake Cook Road
#> 6           Sparks      89431            405 S 21st Street