this function parses coordinates from DMS to decimal degrees

this function parses coordinates from DMS to decimal degrees

Usage

dms2dd(x, lat, long, dms)

Arguments

x

an input data frame to apply operations too

lat

a name of the column holding the latitude values

long

a name of the colymn holding the longitude values

dms

are coordinates in degrees minutes seconds? TRUE for yes, FALSE for decimal degrees

Value

data frame(/tibble) with coordinates unambiguously labeled as being in both degrees, minutes, seconds (_dms) and decimal degrees (_dd).

Examples

 coords <- data.frame(
  longitude_dd = runif(15, min = -120, max = -100),
  latitude_dd = runif(15, min = 35, max = 48)
)
coords_formatted <- dms2dd( coords )
head(coords_formatted)
#>    longitude_dd latitude_dd latitude_dms longitude_dms
#>           <num>       <num>       <char>        <char>
#> 1:   -119.59870    47.96190   N 47°57'43   W 119°35'55
#> 2:   -112.46058    36.93746   N 36°56'15   W 112°27'38
#> 3:   -108.80174    41.74124   N 41°44'28    W 108°48'6
#> 4:   -102.85833    45.99956   N 45°59'58   W 102°51'30
#> 5:   -112.30381    44.33751   N 44°20'15   W 112°18'14
#> 6:   -109.44166    38.13708    N 38°8'13   W 109°26'30
colnames(coords_formatted)
#> [1] "longitude_dd"  "latitude_dd"   "latitude_dms"  "longitude_dms"